requirements.txt
flask
werkzeug
pillow
google-generativeai
python-dotenv
